pyproject.toml
[project]
name = "django-forums-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5",
    "crispy-bootstrap4>=2022.1",
    "dj-database-url>=2.1,<3",
    "django>=4.2,<5.0",
    "django-allauth>=0.61,<0.64",
    "django-cors-headers>=4.0",
    "django-crispy-forms>=2.1",
    "django-debug-toolbar>=4.2",
    "django-lifecycle>=1.0",
    "django-permissions-auditor>=1.2",
    "django-redis>=5.4",
    "djangorestframework>=3.14",
    "gunicorn>=23",
    "martor>=1.6.40,<1.7",
    "psycopg2-binary>=2.9.10",
    "pyjwt[crypto]>=2.8",
    "redis>=4.6",
    "requests>=2.31",
    "requests-oauthlib>=1.3",
    "sentry-sdk>=1.40",
    "whitenoise>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-cov>=5.0",
    "pytest-django>=4.8",
    "pytest-xdist>=3.5",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
norecursedirs = [".*", "venv", ".venv"]
addopts = "-n auto"

[tool.coverage.run]
omit = [
    "*apps.py",
    "project/settings/*",
    "*urls.py",
    "*wsgi.py",
    "*asgi.py",
    "manage.py",
    "*migrations/*",
    "*tests/*",
    "forums/tests.py",
    "pages/*",
    "users/tests.py",
]
branch = true
